Restore `canvas::Frame` API
